Make case sensitivity user settable
The construction of a static decision to treat everything as case sensitive
or not is an assumption that doesn't always hold. Keep the current
behavior, but make it possible for the caller to explicitly set
case sensitivity.

(cherry picked from commit cad87275995dddfd3b59e906a953e0cc4dc8af3d)